move Symfony related Twig variable collector into its own package
