+++ src/fr/adrienbrault/idea/symfony2plugin/templating/variable/collector/GlobalExtensionVariableCollector.java
@@ -9,6 +9,7 @@ import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigUtil;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.TwigFileVariableCollector;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.TwigFileVariableCollectorParameter;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.dict.PsiVariable;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
 import java.util.Map;
@@ -19,7 +20,7 @@ import java.util.Set;
  */
 public class GlobalExtensionVariableCollector implements TwigFileVariableCollector, TwigFileVariableCollector.TwigFileVariableCollectorExt {
     @Override
-    public void collectVars(TwigFileVariableCollectorParameter parameter, Map<String, PsiVariable> variables) {
+    public void collectVars(@NotNull TwigFileVariableCollectorParameter parameter, @NotNull Map<String, PsiVariable> variables) {
         for(PhpClass phpClass : TwigUtil.getTwigExtensionClasses(parameter.getProject())) {
             if(!PhpUnitUtil.isPhpUnitTestFile(phpClass.getContainingFile())) {
                 Method method = phpClass.findMethodByName("getGlobals");
