+++ src/fr/adrienbrault/idea/symfony2plugin/templating/variable/collector/StaticVariableCollector.java
@@ -2,6 +2,7 @@ package fr.adrienbrault.idea.symfony2plugin.templating.variable.collector;
 
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.TwigFileVariableCollector;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.TwigFileVariableCollectorParameter;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.*;
 
@@ -9,10 +10,8 @@ import java.util.*;
  * @author Daniel Espendiller <daniel@espendiller.net>
  */
 public class StaticVariableCollector implements TwigFileVariableCollector {
-
     @Override
-    public void collect(TwigFileVariableCollectorParameter parameter, Map<String, Set<String>> variables) {
+    public void collect(@NotNull TwigFileVariableCollectorParameter parameter, @NotNull Map<String, Set<String>> variables) {
         variables.put("app", new HashSet<>(Collections.singletonList("\\Symfony\\Bundle\\FrameworkBundle\\Templating\\GlobalVariables")));
     }
-
 }
