+++ src/fr/adrienbrault/idea/symfony2plugin/twig/variable/globals/TwigGlobalEnum.java
@@ -1,4 +1,4 @@
-package fr.adrienbrault.idea.symfony2plugin.templating.globals;
+package fr.adrienbrault.idea.symfony2plugin.twig.variable.globals;
 
 /**
  * @author Daniel Espendiller <daniel@espendiller.net>
